Scrumus 38 panel administracyjny

+++ src/main/java/edu/piotrjonski/scrumus/business/IssueManager.java
@@ -5,6 +5,7 @@ import edu.piotrjonski.scrumus.domain.*;
 
 import javax.ejb.Stateless;
 import javax.inject.Inject;
+import java.util.List;
 import java.util.Optional;
 
 @Stateless
@@ -27,6 +28,13 @@ public class IssueManager {
 
     @Inject
     private BacklogDAO backlogDAO;
+
+    @Inject
+    private IssueTypeDAO issueTypeDAO;
+
+    public void deleteIssuesFromProject(String projectKey) {
+        issueDAO.deleteIssuesFromProject(projectKey);
+    }
 
     public Optional<Issue> create(Issue issue, Project project) throws AlreadyExistException, NotExistException {
         if (issueExist(issue)) {
@@ -55,9 +63,12 @@ public class IssueManager {
         return priorityDAO.saveOrUpdate(priority);
     }
 
-    public void deletePriority(Priority priority) throws NotExistException {
+    public void deletePriority(Priority priority) throws NotExistException, IllegalOperationException {
         if (!priorityExist(priority)) {
             throw new NotExistException("Priorytet nie istnieje.");
+        }
+        if (isPriorityInUser(priority)) {
+            throw new IllegalOperationException("Priorytet jest używany przez inne zadania.");
         }
         priorityDAO.delete(priority.getId());
     }
@@ -69,6 +80,30 @@ public class IssueManager {
         return priorityDAO.saveOrUpdate(newPriority);
     }
 
+    public Optional<IssueType> createIssueType(IssueType issueType) throws AlreadyExistException {
+        if (issueTypeExist(issueType)) {
+            throw new AlreadyExistException("Rodzaj zadania już istnieje już istnieje.");
+        }
+        return issueTypeDAO.saveOrUpdate(issueType);
+    }
+
+    public void deleteIssueType(IssueType issueType) throws NotExistException, IllegalOperationException {
+        if (!issueTypeExist(issueType)) {
+            throw new NotExistException("Rodzaj zadania nie istnieje.");
+        }
+        if (isIssueTypeInUse(issueType)) {
+            throw new IllegalOperationException("Rodzaj zadania jest używany przez inne zadania.");
+        }
+        issueTypeDAO.delete(issueType.getId());
+    }
+
+    public Optional<IssueType> updateIssueType(IssueType issueType) throws NotExistException {
+        if (!issueTypeExist(issueType)) {
+            throw new NotExistException("Rodzaj zadania nie istnieje.");
+        }
+        return issueTypeDAO.saveOrUpdate(issueType);
+    }
+
     public Optional<State> createState(State state) throws AlreadyExistException {
         if (stateExist(state)) {
             throw new AlreadyExistException("Stan już istnieje.");
@@ -76,9 +111,12 @@ public class IssueManager {
         return stateDAO.saveOrUpdate(state);
     }
 
-    public void deleteState(State state) throws NotExistException {
+    public void deleteState(State state) throws NotExistException, IllegalOperationException {
         if (!stateExist(state)) {
             throw new NotExistException("Stan nie istnieje.");
+        }
+        if (isStateInUse(state)) {
+            throw new IllegalOperationException("Stan jest używany przez inne zadania.");
         }
         stateDAO.delete(state.getId());
     }
@@ -89,6 +127,46 @@ public class IssueManager {
         }
         return stateDAO.saveOrUpdate(state);
     }
+
+    public List<String> findAllIssueTypeNames() {
+        return issueTypeDAO.findAllNames();
+    }
+
+    public List<String> findAllStateNames() {
+        return stateDAO.findAllNames();
+    }
+
+    public List<String> findAllPriorityNames() {return priorityDAO.findAllNames();}
+
+    public List<IssueType> findAllIssueTypes() {
+        return issueTypeDAO.findAll();
+    }
+
+    public List<State> findAllStates() {return stateDAO.findAll();}
+
+    public List<Priority> findAllPriorities() {return priorityDAO.findAll();}
+
+    public Optional<IssueType> findIssueType(final int issueTypeId) {
+        return issueTypeDAO.findById(issueTypeId);
+    }
+
+    public Optional<State> findState(final int stateId) {
+        return stateDAO.findById(stateId);
+    }
+
+    public Optional<Priority> findPriority(final int priorityId) {
+        return priorityDAO.findById(priorityId);
+    }
+
+    private boolean isIssueTypeInUse(final IssueType issueType) {
+        return issueDAO.isIssueTypeInUse(issueType.getName());
+    }
+
+    private boolean isStateInUse(final State state) {
+        return issueDAO.isStateInUse(state.getName());
+    }
+
+    private boolean isPriorityInUser(final Priority priority) {return issueDAO.isPriorityInUse(priority.getName());}
 
     private void addIssueToBacklog(final Issue issue, Project project) {
         Backlog backlog = backlogDAO.findBacklogForProject(project.getKey())
@@ -116,6 +194,10 @@ public class IssueManager {
 
     private boolean stateExist(final State state) {
         return stateDAO.exist(state.getId());
+    }
+
+    private boolean issueTypeExist(final IssueType issueType) {
+        return issueTypeDAO.exist(issueType.getId());
     }
 
 }
